sender/requirements.txt
aio-pika==7.1.0
Jinja2==3.1.2
pydantic
pydantic-settings==2.0.3
psycopg2-binary==2.9.4